% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_any_to_csv_iso}
\alias{convert_any_to_csv_iso}
\title{Convert any file to csv format}
\usage{
convert_any_to_csv_iso(input_dir, output_dir = NULL)
}
\arguments{
\item{input_dir}{directory address to folder containing files to convert to CSV format}

\item{output_dir}{text in quotes for directory to hold converted CSV files}
}
\value{
CSV files
}
\description{
Convert any file to csv format
}
\examples{
\dontrun{convert_any_to_csv_iso(getwd())}
}
